[LMS] 회원 정보 수정, 회원 탈퇴 경로 설정
DETAIL : 220412
회원 정보 수정, 회원 탈퇴 경로 설정
