+++ teamproject/src/main/resources/templates/member/modifyMember.html
@@ -8,7 +8,7 @@
 	</th:block>
 	
 	<th:block layout:fragment="customContents">
-		<form th:action="@{/member/modifyMember}" method="post">
+		<form th:action="@{/mypage/modifyMember}" method="post">
 			<table border="1">
 				<tbody>
 					<tr>
