+++ teamproject/src/main/resources/templates/member/removeMember.html
@@ -29,7 +29,7 @@
 			}
 		</style>
 		
-		<form id="removeForm" th:action="@{/member/removeMember}" method="post">
+		<form id="removeForm" th:action="@{/mypage/removeMember}" method="post">
 			<table border="1">
 				<tbody>
 					<tr>
